Starting to implement viewport for hardware scrolling.

+++ api/glut/retapif.h
@@ -271,7 +271,7 @@ void retroflat_destroy_bitmap( struct RETROFLAT_BITMAP* bmp ) {
 
 void retroflat_blit_bitmap(
    struct RETROFLAT_BITMAP* target, struct RETROFLAT_BITMAP* src,
-   size_t s_x, size_t s_y, size_t d_x, size_t d_y, size_t w, size_t h,
+   size_t s_x, size_t s_y, int16_t d_x, int16_t d_y, size_t w, size_t h,
    int16_t instance
 ) {
    assert( NULL != src );
